Feature: Automatic update of gh-pages
Changes:
- Git config changes should happen in the copied folder before push

+++ ci/release_documentation.sh
@@ -1,26 +1,20 @@
 #!/bin/bash
 
-# If in docker environment
-if [ -f /.dockerenv ]; then
-  echo "In docker environment"
-  git config --global --add safe.directory /home/user/workdir
-fi
-
 # Check if folder exists
 if [ ! -d "./doc/_site" ]; then
   echo "Documentation does not exist. Exiting."
   exit 1
 fi
 
-# Automated release process credentials
-git config --global user.email "<>"
-git config --global user.name "Documentation release Bot (via GitHub actions)"
-
 # Move the whole directory to a temporary location
 rm -rf /tmp/gh-pages
 rm -rf /tmp/_site
 cp -r . /tmp/gh-pages
 cd /tmp/gh-pages
+
+# Automated release process credentials
+git config --global user.email "<>"
+git config --global user.name "Documentation release Bot (via GitHub actions)"
 
 # Move all the files inside the doc/_site folder to the root directory
 mv ./doc/_site /tmp/
@@ -29,6 +23,12 @@ mv ./doc/_site /tmp/
 git restore .
 git clean -fd
 git checkout gh-pages
+
+# Fail if branch is not gh-pages
+if [ "$(git branch --show-current)" != "gh-pages" ]; then
+  echo "Not on gh-pages branch. Exiting."
+  exit 1
+fi
 
 # Clean out the old documentation
 rm -rf *
